requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ase==3.24.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
biopandas==0.5.1
biopython==1.84
bioservices==1.11.2
bleach==6.2.0
botorch==0.12.0
cachetools==5.5.2
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cpdb-protein==0.2.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
deepdiff==8.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dftd3==1.2.1
dill==0.3.9
distlib==0.3.9
docutils==0.21.2
easydev==0.13.3
easydict==1.13
executing==2.1.0
fastjsonschema==2.21.1
fica==0.4.1
filelock==3.16.1
filtration_domination==0.1.0
flake8==7.1.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gauche==0.1.6
geometric==1.0.2
gevent==24.11.1
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.45
gmsh==4.13.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gpytorch==1.13
GraKeL==0.1.10
graphein==1.7.7
graphviz==0.20.3
greenlet==3.1.1
grequests==0.7.0
gudhi==3.10.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imbalanced-learn==0.13.0
imblearn==0.0
imgcat==0.6.0
iniconfig==2.0.0
ipykernel==6.29.5
ipylab==1.0.0
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.4.38
jaxlib==0.4.38
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupytext==1.16.6
keopscore==2.3
kiwisolver==1.4.8
latexcodec==3.0.0
lazy_loader==0.4
line_profiler==4.2.0
linear-operator==0.5.3
loguru==0.7.3
looseversion==1.1.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.0
ml_dtypes==0.5.1
mmtf-python==1.1.3
monty==2025.1.9
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipers==2.3.2
multipledispatch==1.0.0
mypy==1.14.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numpy==2.3.1
oauthlib==3.3.1
opt_einsum==3.4.0
ordered-set==4.1.0
orderly-set==5.2.3
otter-grader==6.0.4
overrides==7.7.0
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
periodictable==2.0.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
plotly==5.24.1
pluggy==1.5.0
POT==0.9.5
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.0
pybtex==0.24.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
pyfiglet==1.0.2
pyflakes==3.2.0
Pygments==2.19.1
pykeops==2.3
pymatgen==2025.1.9
pyparsing==3.2.1
pyro-api==0.1.2
pyro-ppl==1.9.1
pyscf==2.8.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-on-whales==0.75.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2024.9.4
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-click==1.8.5
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
selfies==2.1.2
Send2Trash==1.8.3
sentry-sdk==2.33.2
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spglib==2.5.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
stack-data==0.6.3
suds-community==1.2.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tinycss2==1.4.0
TopoNetX==0.2.0
torch==2.5.1
torch_geometric==2.5.2
torchvision==0.17.2
torchviz==0.0.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
typeguard==4.4.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uncertainties==3.2.2
uri-template==1.3.0
uritemplate==4.2.0
url-normalize==1.4.3
urllib3==2.3.0
virtualenv==20.28.1
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wget==3.2
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xarray==2025.1.1
xlsxwriter==3.2.5
xmltodict==0.14.2
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
